# data_aggregator_service/data-aggregator-service-openapi.yml
openapi: 3.0.3
info:
  title: Data Aggregator Service API
  description: |
    Микросервис для агрегации данных о перемещениях объектов, генерации отчетов и аналитики.
    Периодически собирает данные от positioning-service, агрегирует их и предоставляет аналитические отчеты.
  version: 1.0.0
  contact:
    name: Команда разработки
    email: analytics@rtls.example.com
servers:
  - url: http://localhost:8085/api/v1
    description: Локальный сервер разработки
  - url: https://analytics.rtls.example.com/api/v1
    description: Продуктивный сервер
...
paths:
...
  /reports/zone-occupancy:
    get:
      tags:
        - Reports
      summary: Отчет по посещаемости зон
      description: Возвращает данные о посещаемости различных зон за указанный период
      operationId: getZoneOccupancyReport
      parameters:
        - name: start_time
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: end_time
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: zone_ids
          in: query
          required: false
          schema:
            type: array
            items:
              type: string
        - name: entity_types
          in: query
          required: false
          schema:
            type: array
            items:
              type: string
              enum: [employee, equipment]
      responses:
        '200':
          description: Успешный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ZoneOccupancyReport'
        '400':
          description: Некорректные параметры запроса
  /reports/time-in-zone:
    get:
      tags:
        - Reports
      summary: Отчет по времени пребывания в зонах
      description: Возвращает данные о времени, проведенном сущностями в различных зонах
      operationId: getTimeInZoneReport
      parameters:
        - name: entity_id
          in: query
          required: false
          schema:
            type: string
        - name: zone_id
          in: query
          required: false
          schema:
            type: string
        - name: start_time
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: end_time
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: group_by
          in: query
          required: false
          schema:
            type: string
            enum: [hour, day, week, month]
            default: day
      responses:
        '200':
          description: Успешный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimeInZoneReport'
        '400':
          description: Некорректные параметры запроса
  /reports/workflow-efficiency:
    get:
      tags:
        - Reports
      summary: Отчет по эффективности рабочих зон
      description: Анализ эффективности использования рабочих зон и маршрутов
      operationId: getWorkflowEfficiencyReport
      parameters:
        - name: start_time
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: end_time
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: zone_ids
          in: query
          required: false
          schema:
            type: array
            items:
              type: string
        - name: entity_ids
          in: query
          required: false
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: Успешный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowEfficiencyReport'
        '400':
          description: Некорректные параметры запроса
...
components:
  schemas:
    ZoneOccupancyReport:
      type: object
      properties:
        report_id:
          type: string
        generated_at:
          type: string
          format: date-time
        period:
          type: object
          properties:
            start_time:
              type: string
              format: date-time
            end_time:
              type: string
              format: date-time
        zones:
          type: array
          items:
            type: object
            properties:
              zone_id:
                type: string
              zone_name:
                type: string
              zone_type:
                type: string
              total_visits:
                type: integer
              unique_entities:
                type: integer
              avg_duration_minutes:
                type: number
              peak_hour:
                type: integer
              hourly_distribution:
                type: object
                additionalProperties:
                  type: integer
              entity_breakdown:
                type: object
                properties:
                  employees:
                    type: integer
                  equipment:
                    type: integer
    TimeInZoneReport:
      type: object
      properties:
        report_id:
          type: string
        generated_at:
          type: string
          format: date-time
        period:
          type: object
          properties:
            start_time:
              type: string
              format: date-time
            end_time:
              type: string
              format: date-time
        group_by:
          type: string
        data:
          type: array
          items:
            type: object
            properties:
              entity_id:
                type: string
              entity_name:
                type: string
              entity_type:
                type: string
              zone_id:
                type: string
              zone_name:
                type: string
              total_time_minutes:
                type: number
              visits_count:
                type: integer
              avg_visit_duration:
                type: number
              first_entry:
                type: string
                format: date-time
              last_exit:
                type: string
                format: date-time
    WorkflowEfficiencyReport:
      type: object
      properties:
        report_id:
          type: string
        generated_at:
          type: string
          format: date-time
        period:
          type: object
          properties:
            start_time:
              type: string
              format: date-time
            end_time:
              type: string
            format: date-time
        zones:
          type: array
          items:
            type: object
            properties:
              zone_id:
                type: string
              zone_name:
                type: string
              utilization_rate:
                type: number
                description: Доля времени, когда зона использовалась
              avg_entities_per_hour:
                type: number
              bottleneck_score:
                type: number
                minimum: 0
                maximum: 1
                description: Оценка узкого места (0 - нет, 1 - критическое)
              peak_hours:
                type: array
                items:
                  type: integer
              workflow_metrics:
                type: object
                properties:
                  avg_transition_time:
                    type: number
                    description: Среднее время перехода между зонами
                  common_routes:
                    type: array
                    items:
                      type: object
                      properties:
                        route:
                          type: array
                          items:
                            type: string
                        frequency:
                          type: integer
...